# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Aplicativo Web para la Centralización de Documentos Electrónicos Tributarios",
        "contact": {
            "name": "Rafael Larrea",
            "email": "jrafael1108@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Autorización"
                ],
                "summary": "Verifica si el usuario tiene acceso al sistema retornando un token de acceso",
                "operationId": "222b2625838e3c57e017987f73598fda",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "dni": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "name": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "email": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "pwd": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/refresh": {
            "post": {
                "tags": [
                    "Autorización"
                ],
                "summary": "Actualiza el token de acceso utilizando un token de actualización válido.",
                "operationId": "refresh",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "refresh_token": {
                                        "description": "El token de actualización",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Token": []
                    },
                    {
                        "Session": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "APIResponse": {
                "description": "",
                "properties": {
                    "Success": {
                        "type": "boolean"
                    },
                    "Message": {
                        "type": "string"
                    },
                    "Data": {
                        "type": "object"
                    },
                    "Exception": {
                        "type": "object"
                    },
                    "Code": {
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "Token": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "bearerFormat": "JWT",
                "in": "header"
            },
            "Session": {
                "type": "apiKey",
                "description": "Enter your session id",
                "name": "SessionId",
                "in": "header"
            },
            "Apikey": {
                "type": "apiKey",
                "description": "Enter your APIKEY for company",
                "name": "APIKEY",
                "in": "header"
            }
        }
    },
...
}